Automation for Labyrinth of Shadows
