+++ src/data/defaults.txt
@@ -3085,6 +3085,7 @@ user	choiceAdventure1488	0
 user	choiceAdventure1489	3
 user	choiceAdventure1491	0
 user	choiceAdventure1494	0
+user	choiceAdventure1499	0
 user	lockedItem4637	true
 user	lockedItem4638	true
 user	lockedItem4639	true
