+++ src/net/sourceforge/kolmafia/session/RufusManager.java
@@ -1,5 +1,8 @@
 package net.sourceforge.kolmafia.session;
 
+import java.util.EnumSet;
+import java.util.HashMap;
+import java.util.Map;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import net.sourceforge.kolmafia.objectpool.ItemPool;
@@ -7,6 +10,7 @@ import net.sourceforge.kolmafia.persistence.ItemDatabase;
 import net.sourceforge.kolmafia.persistence.QuestDatabase;
 import net.sourceforge.kolmafia.persistence.QuestDatabase.Quest;
 import net.sourceforge.kolmafia.preferences.Preferences;
+import net.sourceforge.kolmafia.utilities.ChoiceUtilities;
 
 public class RufusManager {
   private RufusManager() {}
@@ -190,5 +194,286 @@ public class RufusManager {
     }
     // This should not be possible
     return QuestDatabase.UNSTARTED;
+  }
+
+  // Support for The Shadow Labyrinth.
+  //
+  // This is a (turn-free) scheduled choice adventure in the Shadow Rift
+  // It appears even if you don't have a quest from Rufus, but is
+  // modified by such.
+  //
+  // (no quest) - unmodified
+  // entity - replaced by a boss fight
+  // artifact - one of the options is replaced by the artifact
+  // items - unmodified
+  //
+  // Support for it is a fair chunk of code, so it's here, rather than
+  // in the massive ChoiceAdventures package.
+
+  public enum ShadowTheme {
+    FIRE("90-100 Muscle substats", "shadow lighter"),
+    MATH("90-100 Mysticality substats", "shadow heptahedron"),
+    WATER("90-100 Moxie substats", "shadow bucket"),
+    TIME("+3 turns to 3 random effects"),
+    BLOOD("30 Shadow's Heart: Maximum HP +300%", "shadow heart"),
+    COLD("30 Shadow's Chill: Maximum MP +300%", "shadow snowflake"),
+    GHOST("30 Shadow's Thickness: Superhuman (+5) Spooky, Hot, Sleaze resistance", "shadow wave");
+
+    // Class fields
+    final String normal;
+    final String artifact;
+
+    // Lookups for Shadow Themes
+    private static final Map<String, ShadowTheme> decisionToTheme = new HashMap<>();
+    private static final Map<String, ShadowTheme> artifactToTheme = new HashMap<>();
+
+    ShadowTheme(String normal) {
+      this.normal = normal;
+      this.artifact = null;
+    }
+
+    ShadowTheme(String normal, String artifact) {
+      this.normal = normal;
+      this.artifact = artifact;
+    }
+
+    public String getNormal() {
+      return this.normal;
+    }
+
+    public String getArtifact() {
+      return this.artifact;
+    }
+
+    public void populateMaps() {
+      ShadowTheme.decisionToTheme.put(String.valueOf(this.ordinal() + 1), this);
+      if (artifact != null) {
+        ShadowTheme.artifactToTheme.put(this.artifact, this);
+      }
+    }
+
+    public static ShadowTheme findDecision(String decision) {
+      return ShadowTheme.decisionToTheme.get(decision);
+    }
+
+    public static ShadowTheme findArtifact(String artifact) {
+      return ShadowTheme.artifactToTheme.get(artifact);
+    }
+  }
+
+  static {
+    for (var theme : EnumSet.allOf(ShadowTheme.class)) {
+      theme.populateMaps();
+    }
+  }
+
+  // Adjectives for ShadowThemes
+  private static final Map<String, ShadowTheme> adjectiveToTheme = new HashMap<>();
+
+  private static void addAdjectives(ShadowTheme theme, String... adjectives) {
+    for (String adjective : adjectives) {
+      adjectiveToTheme.put(adjective, theme);
+    }
+  }
+
+  static {
+    addAdjectives(
+        ShadowTheme.FIRE,
+        "blazing",
+        "blistering",
+        "burning",
+        "burnt",
+        "charred",
+        "ember-lit",
+        "flame-choked",
+        "scalded",
+        "scalding",
+        "scorched",
+        "scorching",
+        "seared",
+        "singed",
+        "sizzling",
+        "smoldering",
+        "steaming",
+        "white-hot");
+    addAdjectives(
+        ShadowTheme.MATH,
+        "algebraic",
+        "angular",
+        "binomial",
+        "boolean",
+        "Cartesian",
+        "cubic",
+        "decimal",
+        "divided",
+        "Euclidean",
+        "exponential",
+        "Fibonacci",
+        "fractal",
+        "fractional",
+        "geometric",
+        "hyperbolic",
+        "integer",
+        "irrational",
+        "logarithmic",
+        "monomial",
+        "multiplicative",
+        "ordinal",
+        "parabolic",
+        "periodic",
+        "prime",
+        "Pythagorean",
+        "quadratic",
+        "Riemannian",
+        "self-referential",
+        "sinusoidal",
+        "trigonometric",
+        "vector");
+    addAdjectives(
+        ShadowTheme.WATER,
+        "aqueous",
+        "damp",
+        "drenched",
+        "dripping",
+        "drowned",
+        "drowning",
+        "foggy",
+        "humid",
+        "moist",
+        "runny",
+        "soaked",
+        "sodden",
+        "underwater",
+        "wet",
+        "water-logged",
+        "watery");
+    addAdjectives(
+        ShadowTheme.TIME,
+        "ancient",
+        "antique",
+        "broken-down",
+        "crumbling",
+        "decaying",
+        "derelict",
+        "dilapidated",
+        "old",
+        "ramshackle",
+        "rickety",
+        "ruined",
+        "shabby",
+        "unkempt");
+    addAdjectives(
+        ShadowTheme.BLOOD,
+        "bleeding",
+        "blood-drenched",
+        "blood-soaked",
+        "bloodstained",
+        "bloody",
+        "crimson",
+        "hematic",
+        "pulsing",
+        "sanguine",
+        "vein-shot",
+        "veiny");
+    addAdjectives(
+        ShadowTheme.COLD,
+        "arctic",
+        "chilly",
+        "cold-numbed",
+        "freezing",
+        "frigid",
+        "frost-rimed",
+        "frosty",
+        "frozen",
+        "hyperborean",
+        "iced-over",
+        "icy",
+        "snow-covered",
+        "spectral",
+        "wintry");
+    addAdjectives(
+        ShadowTheme.GHOST,
+        "diaphanous",
+        "ephemeral",
+        "ghostly",
+        "gossamer",
+        "half-there",
+        "insubstantial",
+        // The only "two word" adjective
+        "nearly invisible",
+        "see-through",
+        "translucent",
+        "transparent",
+        "wispy");
+  }
+
+  public static ShadowTheme shadowLabyrinthTheme(String text) {
+    // VERB the ADJECTIVE NOUN
+    //
+    // VERB is an action phrase
+    // ADJECTIVE is a single word - with one known exception
+    // NOUN is a single-word destination
+
+    int the = text.indexOf(" the ");
+    int place = text.lastIndexOf(" ");
+    if (the == -1 || place == -1) {
+      return null;
+    }
+
+    // String verb = text.substring(0, the);
+    String adjective = text.substring(the + 5, place);
+    // String noun = text.substring(place + 1);
+
+    return adjectiveToTheme.get(adjective);
+  }
+
+  public static ChoiceOption shadowLabyrinthSpoiler(String text) {
+    ShadowTheme theme = shadowLabyrinthTheme(text);
+    String spoiler = "unknown theme";
+    if (theme != null) {
+      spoiler = theme.getNormal();
+      if (Preferences.getString("rufusQuestType").equals("artifact")
+          && QuestDatabase.isQuestStep(Quest.RUFUS, QuestDatabase.STARTED)) {
+        String target = Preferences.getString("rufusQuestTarget");
+        String artifact = theme.getArtifact();
+        if (target.equals(artifact)) {
+          spoiler = artifact;
+        }
+      }
+    }
+    return new ChoiceOption(spoiler);
+  }
+
+  // Automation of the Labyrinth of Shadows
+  public static String specialChoiceDecision(String decision, String responseText) {
+    // decisions for choiceAdventure1499:
+    // 0 - show in browser
+    // 1-7 - correspond to ShadowTheme 0-6
+
+    boolean artifact =
+        QuestDatabase.isQuestStep(Quest.RUFUS, QuestDatabase.STARTED)
+            && Preferences.getString("rufusQuestType").equals("artifact");
+    ShadowTheme needed =
+        artifact
+            ? ShadowTheme.findArtifact(Preferences.getString("rufusQuestTarget"))
+            : ShadowTheme.findDecision(decision);
+    if (needed == null) {
+      // This should not happen. Show in browser
+      return "0";
+    }
+
+    // See what options are on offer in the responseText
+    Map<Integer, String> choices = ChoiceUtilities.parseChoices(responseText);
+    int options = choices.size();
+
+    for (int i = 2; i <= 4; ++i) {
+      ShadowTheme offered = shadowLabyrinthTheme(choices.get(i));
+      if (offered == needed) {
+        return String.valueOf(i);
+      }
+    }
+
+    // The desired theme is not available. Tell KoL to randomize again
+    return "1";
   }
 }
